+++ apps/workflows-dashboard/src/domains/workflow-definitions/workflow-definitions-metrics/workflow-definitions-metrics.types.ts
@@ -1,0 +1,4 @@
+export interface IWorkflowDefinitionByVariantMetric {
+  workflowDefinitionVariant: string;
+  count: number;
+}