+++ apps/workflows-dashboard/src/pages/Filters/types/filters-query-params.ts
@@ -1,0 +1,4 @@
+export interface FiltersPageFilterQuery {
+  page?: number;
+  limit?: number;
+}