+++ apps/workflows-dashboard/src/router.tsx
@@ -1,7 +1,12 @@
 import { App } from '@/App';
 import { withSessionProtected } from '@/common/hocs/withSessionProtected';
+import { AlertDefinitions } from '@/pages/AlertDefinitions';
+import { Filters } from '@/pages/Filters';
 import { Overview } from '@/pages/Overview';
 import { SignIn } from '@/pages/SignIn';
+import { UIDefinitions } from '@/pages/UIDefinitions';
+import { WorkflowDefinition } from '@/pages/WorkflowDefinition';
+import { WorkflowDefinitions } from '@/pages/WorkflowDefinitions';
 import { Workflows } from '@/pages/Workflows';
 import { createBrowserRouter, Navigate } from 'react-router-dom';
 
@@ -20,11 +25,32 @@ export const router = createBrowserRouter([
       },
       {
         path: 'overview',
+        // TODO: get rid of this hook and rework routing to use authenticated layout
         Component: withSessionProtected(Overview),
       },
       {
         path: 'workflows',
         Component: withSessionProtected(Workflows),
+      },
+      {
+        path: 'workflow-definitions',
+        Component: withSessionProtected(WorkflowDefinitions),
+      },
+      {
+        path: 'workflow-definitions/:id',
+        Component: withSessionProtected(WorkflowDefinition),
+      },
+      {
+        path: 'filters',
+        Component: withSessionProtected(Filters),
+      },
+      {
+        path: 'ui-definitions',
+        Component: withSessionProtected(UIDefinitions),
+      },
+      {
+        path: 'alert-definitions',
+        Component: withSessionProtected(AlertDefinitions),
       },
     ],
   },