+++ examples/report-generation-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/report-generation-example
 
+## 0.2.6
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.8
+
+## 0.2.5
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.7
+
 ## 0.2.4
 
 ### Patch Changes